Start to reorganize code (save changes from Friday)
This is a work in progress commit since I didn't get to do all the things I
wanted to do on Friday. On this commit I started to move code around and to
switch over to using SpatialExperiment (from bioc-devel) as well as spatialLIBD
also from bioc-devel.

I'll explain these changes at 1pm