semi-successful tangram run
I ran Tangram on my Nvidia GeForce RTX 3080 and it completed in about 30 minutes. However, many of the plots did not work, and I am also not sure about much of the code. For example, I am not sure if I'm supposed to use the imagerow/imagecolumn or row/column observations in the Visium data.