corrected h5ad conversion code
for some reason we (I) have been writing the modified RSE objects used for marker selections to h5ad instead of the original objects. We decided this might be the reason behind why PCP4 was not showing up in the Tangram step, since it was being filtered out by the `expression_cutoff` portion of the script.