Resolve #23
We mostly adapted the code from
https://github.com/LieberInstitute/HumanPilot/blob/master/Analysis/Layer_Guesses/dlpfc_snRNAseq_annotation.R
and also used functions in spatialLIBD that didn't exist when that script was
made by Andrew.

We only computed the `enrichment` t-statistics (one cluster vs others) and not
the `anova` or `pairwise` ones. Note that the correlation with the layer
`enrichment` t-statistics improved when we kept the top 100 genes per layer (and
WM). The inhibitory neuron don't seem to have much spatial registration.

We should discuss these results with everyone else in the spatial meeting.

Co-authored-by: Nick Eagles <nick.eagles@libd.org>
Co-authored-by: Abby Spangler <aspangle@gmail.com>