Explore a few options for locating low library size spots.
We tried within some BayesSpace cluster 6 from k = 7 (with a lower nmads threshold in scran::isOutlier) in a sample-aware way.

We noticed that cluster 12 from k = 15 was a pretty good cluster for locating the "bad spots".

In the end, we are leaning towards dropping the spots with a low library size from scran::isOutlier() run in a sample-aware way.

Co-authored-by: Abby Spangler <aspangle@gmail.com>
Co-authored-by: Heena Divecha <heenadivecha@gmail.com>